+++ src/components/Map/Layers/Buildings/HelsinkiBuildings/layerConf.ts
@@ -4,66 +4,133 @@ import _ from 'lodash'
 
 import { fillOpacity, roundToSignificantDigits } from '#/utils/mapUtils'
 import { LayerId, LayerConf } from '#/types/map'
-import { layerOptions } from './constants'
 import Popup from './Popup'
-
-const SERVER_URL = process.env.NEXT_PUBLIC_GEOSERVER_URL
 
 const id: LayerId = 'helsinki_buildings'
 
 const getStyle = async (): Promise<MbStyle> => {
-  const sources: any = {}
-  let layers: any = []
+  const sourceNames = ['helsinki_buildings', 'helsinki_puretut']
 
-  for (const layerId in layerOptions) {
-    const options = layerOptions[layerId]
-    sources[layerId] = {
-      type: 'vector',
-      scheme: 'tms',
-      tiles: [`${SERVER_URL}/gwc/service/tms/1.0.0/misc:${options.serverId}@EPSG:900913@pbf/{z}/{x}/{y}.pbf`],
-      minzoom: options.minzoom,
-      maxzoom: options.maxzoom,
-      bounds: [19, 59, 32, 71], // Finland
-      attribution: '<a href="https://www.hel.fi">© City of Helsinki</a>',
-      promoteId: 'id',
-    }
-
-    layers = [
-      ...layers,
+  return {
+    version: 8,
+    name: id,
+    sources: {
+      [sourceNames[0]]: {
+        type: 'vector',
+        tiles: ['https://server.avoin.org/data/map/helsinki-buildings/{z}/{x}/{y}.pbf'],
+        // tiles: ['https://gis.avoin.org/geoserver/gwc/service/tms/1.0.0/misc:helsinkiblockhouses@pbf/{z}/{x}/{y}.pbf'],
+        maxzoom: 14,
+        // Bounds source: https://koordinates.com/layer/4257-finland-11000000-administrative-regions/
+        // select ST_Extent(ST_Transform(ST_SetSRID(geom,3067), 4326))
+        // from "finland-11000000-administrative-regions" where kunta_ni1='Helsinki';
+        bounds: [24, 59, 26, 61],
+        attribution: '<a href="https://www.hel.fi">© City of Helsinki</a>',
+      },
+      [sourceNames[1]]: {
+        type: 'vector',
+        tiles: ['https://server.avoin.org/data/map/hel-puretut/{z}/{x}/{y}.pbf.gz?v=0'],
+        // tiles: ['https://gis.avoin.org/geoserver/gwc/service/tms/1.0.0/misc:helsinkiotherbuildings@pbf/{z}/{x}/{y}.pbf'],
+        maxzoom: 14,
+        // Bounds source: https://koordinates.com/layer/4257-finland-11000000-administrative-regions/
+        // select ST_Extent(ST_Transform(ST_SetSRID(geom,3067), 4326))
+        // from "finland-11000000-administrative-regions" where kunta_ni1='Helsinki';
+        bounds: [24, 59, 26, 61],
+        attribution: '<a href="https://www.hel.fi">© City of Helsinki</a>',
+      },
+    },
+    layers: [
       {
-        id: `${layerId}-fill`,
-        source: layerId,
-        'source-layer': options.serverId,
+        id: sourceNames[0] + '-fill',
+        source: sourceNames[0],
+        'source-layer': 'Rakennukset_alue',
         type: 'fill',
         paint: {
           'fill-color': 'cyan',
-          'fill-opacity': layerId === 'helsinki-buildings' ? 1 : fillOpacity,
+          'fill-opacity': fillOpacity,
         },
         BEFORE: 'FILL',
       },
       {
-        id: `${layerId}-outline`,
-        source: layerId,
-        'source-layer': options.serverId,
+        id: sourceNames[0] + '-outline',
+        source: sourceNames[0],
+        'source-layer': 'Rakennukset_alue',
         type: 'line',
-        minzoom: 0,
+        minzoom: 11,
         paint: {
           'line-opacity': 0.75,
         },
         BEFORE: 'OUTLINE',
-      },     
-    ]
-  }
-  return {
-    version: 8,
-    name: id,
-    sources: sources,
-    layers: layers,
+      },
+      {
+        id: sourceNames[0] + '-co2',
+        source: sourceNames[0],
+        'source-layer': 'Rakennukset_alue',
+        type: 'symbol',
+        minzoom: 16,
+        paint: {},
+        layout: {
+          'symbol-placement': 'point',
+          'text-font': ['Open Sans Regular'],
+          'text-size': 20,
+          'text-field': [
+            'case',
+            ['has', 'i_raktilav'],
+            [
+              'let',
+              'co2',
+              ['/', ['*', 15, ['to-number', ['get', 'i_raktilav'], 0]], 1000],
+              [
+                'concat',
+                roundToSignificantDigits(2, ['var', 'co2']), // kg -> tons
+                ' t CO2e/y',
+              ],
+            ],
+            '',
+          ],
+        },
+        BEFORE: 'LABEL',
+      },
+      {
+        id: sourceNames[1] + '-fill',
+        source: sourceNames[1],
+        'source-layer': 'default',
+        type: 'fill',
+        paint: {
+          'fill-color': 'black',
+          'fill-opacity': fillOpacity,
+        },
+        BEFORE: 'FILL',
+      },
+      {
+        id: sourceNames[1] + '-outline',
+        source: sourceNames[1],
+        'source-layer': 'default',
+        type: 'line',
+        minzoom: 11,
+        paint: {
+          'line-opacity': 0.75,
+        },
+        BEFORE: 'OUTLINE',
+      },
+      {
+        id: sourceNames[1] + '-sym',
+        source: sourceNames[1],
+        'source-layer': 'default',
+        type: 'symbol',
+        minzoom: 16,
+        paint: {},
+        layout: {
+          'symbol-placement': 'point',
+          'text-font': ['Open Sans Regular'],
+          'text-size': 20,
+          'text-field': '',
+        },
+        BEFORE: 'LABEL',
+      },
+    ],
   }
 }
 
-const layerConf: LayerConf = { id: id, style: getStyle, popup: Popup, useMb: true, }
+const layerConf: LayerConf = { id: id, style: getStyle, popup: Popup }
 
 export default layerConf
-
-
